Address compiler warnings

+++ app/src/main/java/com/kylecorry/trail_sense/tools/tools/ui/ToolsSettingsFragment.kt
@@ -52,7 +52,7 @@ class ToolsSettingsFragment : AndromedaPreferenceFragment() {
                         val selected = ids[it]
                         bottomNavSelection =
                             prefs.bottomNavigationTools.mapIndexed { index, id ->
-                                if (index == i) selected.toLong() else id
+                                if (index == i) selected else id
                             }
                         prefs.bottomNavigationTools = bottomNavSelection
                         preference.summary = names[it]
